UserMetaField extends I18nAbstractFormField, fixed tab malfunction and design issues in inspect and site_settings page, applied quick fix for language ordering issue by switching en and zh-hant in base.py
